[FLINK-9787][core] Let getGlobalJobParameters never return null
When accessing ExecutionConfig#getGlobalJobParameters previously it
would return null if no instance of GlobalJobParameters was set yet.
With these changes this will return a new instance of GlobalJobParameters
if none are set, which allows for always having a safe call to
.toMap() instead of needing to have null checks everywhere this
is used.

This also removes now unnecessary null-checks from places in
the codebase where they where previously needed